requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data validation and settings
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Database - PostgreSQL (using asyncpg for better compatibility)
sqlalchemy>=2.0.25
asyncpg>=0.29.0

# Database - Vector Store
qdrant-client>=1.7.3

# Google Gemini API (for chatbot)
google-generativeai>=0.3.0

# Google Translate (for translation - free, no API key needed)
deep-translator>=1.11.4

# HTTP client
httpx>=0.26.0

# Environment variables
python-dotenv>=1.0.0

# CORS middleware
python-multipart>=0.0.6

# Authentication
bcrypt>=4.0.0
pyjwt>=2.8.0

# Testing (optional, for development)
pytest>=7.4.4
pytest-asyncio>=0.23.3
